+++ packages/util-crypto/src/blake2/asU8a.ts
@@ -5,7 +5,7 @@ import type { HexString } from '@polkadot/util/types';
 
 import { hasBigInt, u8aToU8a } from '@polkadot/util';
 import { blake2b, isReady } from '@polkadot/wasm-crypto';
-import { blake2b as blake2bJs } from '@polkadot/x-noble-hashes/blake2b';
+import { blake2b as blake2bJs } from '@noble/hashes/lib/blake2b';
 
 import { createAsHex } from '../helpers';