+++ packages/util-crypto/src/scrypt/encode.ts
@@ -6,7 +6,7 @@ import type { Params } from './types';
 
 import { hasBigInt, objectSpread, u8aToU8a } from '@polkadot/util';
 import { isReady, scrypt } from '@polkadot/wasm-crypto';
-import { scrypt as scryptJs } from '@polkadot/x-noble-hashes/scrypt';
+import { scrypt as scryptJs } from '@noble/hashes/lib/scrypt';
 
 import { randomAsU8a } from '../random/asU8a';
 import { DEFAULT_PARAMS } from './defaults';