+++ packages/util-crypto/src/sha/asU8a.ts
@@ -5,8 +5,8 @@ import type { HexString } from '@polkadot/util/types';
 
 import { hasBigInt, u8aToU8a } from '@polkadot/util';
 import { isReady, sha256, sha512 } from '@polkadot/wasm-crypto';
-import { sha256 as sha256Js } from '@polkadot/x-noble-hashes/sha256';
-import { sha512 as sha512Js } from '@polkadot/x-noble-hashes/sha512';
+import { sha256 as sha256Js } from '@noble/hashes/lib/sha256';
+import { sha512 as sha512Js } from '@noble/hashes/lib/sha512';
 
 import { createBitHasher } from '../helpers';